Move module-level functions into *_helper.py files
