+++ tests/core_tests.py
@@ -34,6 +34,7 @@ import pandas as pd
 import psycopg2
 import sqlalchemy as sqla
 
+from superset.views.core_helpers import deserialize_results_payload
 from tests.test_app import app
 from superset import dataframe, db, jinja_context, security_manager, sql_lab
 from superset.connectors.sqla.models import SqlaTable
@@ -899,7 +900,7 @@ class CoreTests(SupersetTestCase):
         self.assertIsInstance(serialized_payload, str)
 
         query_mock = mock.Mock()
-        deserialized_payload = views._deserialize_results_payload(
+        deserialized_payload = deserialize_results_payload(
             serialized_payload, query_mock, use_new_deserialization
         )
 
@@ -947,7 +948,7 @@ class CoreTests(SupersetTestCase):
             query_mock = mock.Mock()
             query_mock.database.db_engine_spec.expand_data = expand_data
 
-            deserialized_payload = views._deserialize_results_payload(
+            deserialized_payload = deserialize_results_payload(
                 serialized_payload, query_mock, use_new_deserialization
             )
             payload["data"] = dataframe.SupersetDataFrame.format_data(cdf.raw_df)
